refactor: split tests into individual modules by type
