+++ secure_logger/tests/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""Init file for tests module."""
+import sys
+
+
+sys.path.append("../")  # noqa: E402
